# openapi.yaml
openapi: 3.0.0
info:
    title: TraderJoe API
    description: Easily retrieve info on the state of Trader Joe.
    version: 0.0.1
basePath: /v1
servers:
    - url: https://api.traderjoexyz.com/v1
      description: Production
paths:
...
    /supply/user/{userAddress}:
        get:
            summary: Outstanding supply of user
            tags:
                - bank
            parameters:
                - name: userAddress
                  in: path
                  required: true
                  description: Address of user
                  schema:
                      type: string
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: string
                                        example: '30009819328012'
                                    status:
                                        $ref: '#/components/schemas/status'
    /borrow/user/{userAddress}:
        get:
            summary: Outstanding borrow of user
            tags:
                - bank
            parameters:
                - name: userAddress
                  in: path
                  required: true
                  description: Address of user
                  schema:
                      type: string
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: string
                                        example: '30009819328012'
                                    status:
                                        $ref: '#/components/schemas/status'
    /{userAddress}/net/apy:
        get:
            summary: Net APY of user with all positions
            tags:
                - bank
            parameters:
                - name: userAddress
                  in: path
                  required: true
                  description: Address of user
                  schema:
                      type: string
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: string
                                        example: '30009819328012'
                                    status:
                                        $ref: '#/components/schemas/status'
...
components:
    schemas:
...
        status:
            type: object
            properties:
                timestamp:
                    type: string
                    example: '1641341612759'
                error_code:
                    type: integer
                    example: 400
                error_message:
                    type: string
                    example: 'Error getting supply'
                error_trace:
                    type: string
                    example: 'at Object.require.extensions.<computed> [as .ts] (D:\Documents\Development\crypto\traderjoe-api\node_modules\ts-node\src\index.ts:1374:12)'
...
    securitySchemes:
        - api_key:
              type: apiKey
              name: api_key
              in: header
        - admin:
              type: http
              scheme: basic